added crud
In this i have created 4 new 5 new files, addcollection.js was just for creating data model in the firebase, crud.js is to perform all the action on the database, firebase.js is for creating a connection b/w firebase DB. index.js will have all the form values we can permorm validations on this then we can transfer that data to crud.

The most important file is serviceAccountKey.json where we have the firebase api key and all the information about db, so do not forward it to anybody.

if you wanna use this just installl fire base into your console globaly

and use npm install firebase-admin

and to try updating data into db you can try index.js, edit the data in that file, and node index.js